+++ Assets/FashionAdSdk/FashionAd.cs
@@ -4,183 +4,71 @@ using UnityEngine;
 using UnityEngine.Networking;
 using Newtonsoft.Json;
 using UnityEngine.UI;
+using Tentuplay.FashionAd;
 public class FashionAd : MonoBehaviour
 {
     static string clientId; // 클라이언트(게임사) 아이디 세팅하기 
     static string userId = "james"; // 유저 아이디 갖고오기
 
     string adId;
-    public AdData adData;
+    public FashionAdApi.AdData panelAdData;
+    public FashionAdApi.AdData popUpAdData;
 
     public Button BuyButton;
-    public PanelShape panelShape;
+    public FashionAdApi.PanelShape panelShape;
 
     public void ShowAd()
     {
-        StartCoroutine(SendShowAd2(userId, panelShape));
+        StartCoroutine(ShowAdChain(userId, panelShape, (x) => panelAdData = x ) ); ;
     }
 
-    public void GoToWebSite()
+    IEnumerator ShowAdChain(string userId, FashionAdApi.PanelShape panelShape, System.Action<FashionAdApi.AdData> var)// 이 부분에서 ADIMPRESSION 할 때 adData 밖으로 빼야 함 api로 메소드들 다 밖으로 뺀다면
     {
-        if (adData.productUrl != null)
+        FashionAdApi fApi = gameObject.AddComponent<FashionAdApi>();
+        yield return StartCoroutine(fApi.sendSetImageAdData(userId, panelShape, var));//먼저 광고 정보를 갖고온다
+        StartCoroutine(fApi.sendShowAd(panelAdData.adSourceUrl));//광고 정보 갖고오면 이미지url넣어서 보여준다.
+        fApi.AdImpression(userId, panelAdData.adId, panelAdData.ownerId);
+    }
+
+    /*
+     광고판에 광고 클릭 후 돌아왔을 때 어울리는 옷 팝업광고 보여주기
+     */
+    public void ShowPopUpAd()
+    {
+        StartCoroutine(ShowPopUpAdChain());
+    }
+
+    IEnumerator ShowPopUpAdChain()
+    {
+        FashionAdApi fApi = gameObject.AddComponent<FashionAdApi>();
+
+        yield return StartCoroutine(fApi.SetPopUpAdData(userId, "2", (x) => popUpAdData = x));
+        StartCoroutine(fApi.ShowPopUpAd(popUpAdData.adSourceUrl, popUpAdData, GoPopUpAdSite));
+        fApi.AdImpression(userId, popUpAdData.adId, popUpAdData.ownerId);
+    }
+
+
+    public void GoPanelAdSite()
+    {
+        FashionAdApi fApi = gameObject.AddComponent<FashionAdApi>();
+        if (panelAdData.productUrl != null)
         {
-            string url = adData.productUrl;
+            string url = panelAdData.productUrl;
             Application.OpenURL(url);
-            AdClick(userId, adData.adId, adData.ownerId);
+            fApi.AdClick(userId, panelAdData.adId, panelAdData.ownerId);
         }
     }
 
-    public IEnumerator SendShowAd2(string userId, PanelShape panelShape)// 이 부분에서 ADIMPRESSION 할 때 adData 밖으로 빼야 함 api로 메소드들 다 밖으로 뺀다면
+    public void GoPopUpAdSite()
     {
-        yield return StartCoroutine(sendSetImageAdData(userId, panelShape));//먼저 광고 정보를 갖고온다
-        StartCoroutine(sendShowAd(adData.adSourceUrl));//광고 정보 갖고오면 이미지url넣어서 보여준다.
-        AdImpression(userId, adData.adId, adData.ownerId);
-    }
-
-    IEnumerator sendShowAd(string url)
-    {
-        UnityWebRequest www = UnityWebRequestTexture.GetTexture(url);
-        yield return www.SendWebRequest();
-
-        if (www.isNetworkError || www.isHttpError)
+        FashionAdApi fApi = gameObject.AddComponent<FashionAdApi>();
+        if (popUpAdData.productUrl != null)
         {
-            Debug.Log(www.error);
-        }
-        else
-        {
-            this.GetComponent<RawImage>().texture = ((DownloadHandlerTexture)www.downloadHandler).texture;
+            string url = popUpAdData.productUrl;
+            Application.OpenURL(url);
+            fApi.AdClick(userId, popUpAdData.adId, popUpAdData.ownerId);
         }
     }
 
-    public void AdClick(string userId, string adId, string ownerId)
-    {
-        StartCoroutine(sendAdClick(userId, adId, ownerId));
-    }
-    public void AdImpression(string userId, string adId, string ownerId)
-    {
-        StartCoroutine(sendAdImpression(userId, adId, ownerId));
-    }
-    //보조 메소드
-    private IEnumerator sendAdClick(string userId, string adId, string ownerId)
-    {
-        List<IMultipartFormSection> formData = new List<IMultipartFormSection>();
-        formData.Add(new MultipartFormDataSection("adId", adId));
-        formData.Add(new MultipartFormDataSection("userId", userId));
-        formData.Add(new MultipartFormDataSection("ownerId", ownerId));
-        formData.Add(new MultipartFormDataSection("type", "adClick"));
-        formData.Add(new MultipartFormDataSection("clientId", "clientId"));
 
-        UnityWebRequest www = UnityWebRequest.Post("http://52.79.143.149/DataProcessing/putAdData.php", formData);
-        yield return www.SendWebRequest();
-
-        if (www.isNetworkError || www.isHttpError)
-        {
-            Debug.Log(www.error);
-        }
-        else
-        {
-            Debug.Log("Form upload complete!");
-            Debug.Log(www.downloadHandler.text);
-        }
-    }
-    private IEnumerator sendAdImpression(string userId, string adId, string ownerId)
-    {
-        List<IMultipartFormSection> formData = new List<IMultipartFormSection>();
-        formData.Add(new MultipartFormDataSection("adId", adId));
-        formData.Add(new MultipartFormDataSection("userId", userId));
-        formData.Add(new MultipartFormDataSection("ownerId", ownerId));
-        formData.Add(new MultipartFormDataSection("type", "adImpression"));
-        formData.Add(new MultipartFormDataSection("clientId", "clientId"));
-
-        UnityWebRequest www = UnityWebRequest.Post("http://52.79.143.149/DataProcessing/putAdData.php", formData);
-        yield return www.SendWebRequest();
-
-        if (www.isNetworkError || www.isHttpError)
-        {
-            Debug.Log(www.error);
-        }
-        else
-        {
-            Debug.Log("Form upload complete!");
-        }
-    }
-    private IEnumerator sendSetImageAdData(string userId, PanelShape adPanelShape)
-    {
-        List<IMultipartFormSection> formData = new List<IMultipartFormSection>();
-        formData.Add(new MultipartFormDataSection("userId", userId));
-        formData.Add(new MultipartFormDataSection("adPanelShape", adPanelShape.ToString()));
-        formData.Add(new MultipartFormDataSection("clientId", "clientId"));
-
-        UnityWebRequest www = UnityWebRequest.Post("http://52.79.143.149/DataProcessing/getAdData.php", formData);
-        yield return www.SendWebRequest();
-
-        if (www.isNetworkError || www.isHttpError)
-        {
-            Debug.Log(www.error);
-        }
-        else
-        {
-            Debug.Log("Form upload complete!");
-            string data = www.downloadHandler.text;
-            Debug.Log(data);
-            adData = JsonConvert.DeserializeObject<AdData>(data);
-        }
-    }
-
-    //보조 enum ,calss
-    public class AdData
-    {
-        public string adId;
-        public string adSourceUrl;
-        public string productUrl;
-        public string ownerId;
-    }
-    public enum Category
-    {
-        man_knit,
-        man_tShirts,
-        man_shirts,
-        man_cardigan,
-        man_jumper,
-        man_jacket,
-        man_coat,
-        man_blueJean,
-        man_pants,
-        man_sweatPants,
-        man_sweatShirts,
-        man_sneakers,
-        man_boots,
-        man_slipper,
-        man_moccasin,
-        man_sportsShoes,
-        man_worker,
-        man_dressShoes,
-        woman_knit,
-        woman_cardigan,
-        woman_onePiece,
-        woman_tShirts,
-        woman_blouse,
-        woman_jumper,
-        woman_jacket,
-        woman_coat,
-        woman_blueJean,
-        woman_skirt,
-        woman_leggings,
-        woman_pants,
-        woman_sweatPants,
-        woman_sweatShirts,
-        woman_boots,
-        woman_worker,
-        woman_heel,
-        woman_sportsShoes,
-        woman_flatShoes,
-        woman_bootie,
-        woman_sneakers,
-    }
-    public enum PanelShape
-    {
-        img11,
-        img43,
-        img34
-    }
 }
